chore: Refactor Gesture Recognition
Refactored the Camera and HandSpeller classes to improve gesture recognition integration and frame processing. The gesture recognizer is now called within the video stream loop, and bounding box drawing is conditional on the capture mode. The Hand and HandSpeller classes were cleaned up for clarity, and the debug button in main.py now launches the camera in live mode instead of using a static imag because all the time is frame who is precessed